+++ lib/code_counter.rb
@@ -1,0 +1,3 @@
+require File.join(File.dirname(__FILE__), 'code_counter', 'version')
+require File.join(File.dirname(__FILE__), 'code_counter', 'engine')
+require File.join(File.dirname(__FILE__), 'code_counter', 'rake')