+++ lib/code_statistics.rb
@@ -1,3 +1,0 @@
-require File.join(File.dirname(__FILE__), 'code_statistics', 'version')
-require File.join(File.dirname(__FILE__), 'code_statistics', 'code_statistics')
-require File.join(File.dirname(__FILE__), 'tasks', 'code_stats')