+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Switch to `rspec`.
 * Begin adding tests for CLI.
 * Fix an uninitialized variable bug.
+* Paths listed in `ADDITIONAL_SOURCE_DIRECTORIES` don't need to be relative to
+  the current directory anymore.
 
 
 ## v0.3.0